refactor encoding guesser
