+++ pkg/handler/processor/guesser/encoding_guesser_test.go
@@ -1,0 +1,72 @@
+//
+// Copyright 2022 The GUAC Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package guesser
+
+import (
+	"testing"
+
+	"github.com/guacsec/guac/internal/testing/testdata"
+	"github.com/guacsec/guac/pkg/handler/processor"
+)
+
+func Test_guessEncoding(t *testing.T) {
+	tests := []struct {
+		name         string
+		doc          processor.Document
+		wantErr      bool
+		wantMimeType string
+	}{
+		{
+			name: "valid .bz2 format document",
+			doc: processor.Document{
+				Blob:   []byte(testdata.CycloneDXBz2Example),
+				Type:   processor.DocumentUnknown,
+				Format: processor.FormatUnknown,
+				SourceInformation: processor.SourceInformation{
+					Collector: "a-collector",
+					Source:    "a-source",
+				},
+			},
+			wantErr:      false,
+			wantMimeType: "application/x-bzip2",
+		},
+		{
+			name: "valid .zst format document",
+			doc: processor.Document{
+				Blob:   []byte(testdata.CycloneDXZstdExample),
+				Type:   processor.DocumentUnknown,
+				Format: processor.FormatUnknown,
+				SourceInformation: processor.SourceInformation{
+					Collector: "a-collector",
+					Source:    "a-source",
+				},
+			},
+			wantErr:      false,
+			wantMimeType: "application/zstd",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			mimeType, err := detectFileEncoding(&tt.doc)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("detectFileEncoding() error = %v, wantErr %v", err, tt.wantErr)
+			}
+			if mimeType != tt.wantMimeType {
+				t.Errorf("detectFileEncoding() incorrect mimeType = %v", mimeType)
+			}
+		})
+	}
+}
