+++ pkg/handler/processor/process/process.go
@@ -16,14 +16,12 @@
 package process
 
 import (
-	"bufio"
 	"bytes"
 	"compress/bzip2"
 	"context"
 	"encoding/xml"
 	"fmt"
 	"io"
-	"net/http"
 	"path/filepath"
 	"strings"
 
@@ -226,22 +224,9 @@ func decodeDocument(ctx context.Context, i *processor.Document) error {
 		if ok {
 			i.Encoding = encoding
 		} else {
-			var mimeType string
-			var bzipMimeType = "application/x-bzip2"
-			var zstdMimeType = "application/zstd"
-			mimeType, err = detectFileEncoding(i)
+			err := guesser.GuessEncoding(ctx, i)
 			if err != nil {
 				return fmt.Errorf("failure while attempting to detect file encoding: %w", err)
-			}
-			switch mimeType {
-			case bzipMimeType:
-				i.Encoding = processor.EncodingBzip2
-			case zstdMimeType:
-				i.Encoding = processor.EncodingZstd
-			default:
-			}
-			if i.Encoding != "" {
-				logger.Infof("byte analysis detected encoding:  %v", i.Encoding)
 			}
 		}
 	}
@@ -272,30 +257,4 @@ func decompressDocument(i *processor.Document, reader io.Reader) error {
 	return nil
 }
 
-func detectFileEncoding(i *processor.Document) (string, error) {
 
-	//create a bufio.Reader so we can 'peek' at the first few bytes without consuming
-	bReader := bytes.NewReader(i.Blob)
-	reader := bufio.NewReader(bReader)
-	blankType := ""
-	if len(i.Blob) >= 16 {
-		testBytes, err := reader.Peek(16)
-		if err != nil {
-			return "", err
-		}
-
-		//find the content mime-type from the first few bytes
-		contentType := http.DetectContentType(i.Blob)
-		//octet-stream is the default meaning that no encoding has been found
-		if contentType == "application/octet-stream" {
-			if testBytes[0] == 0x42 && testBytes[1] == 0x5A && testBytes[2] == 0x68 {
-				return "application/x-bzip2", nil
-			}
-			if testBytes[3] == 0xFD && testBytes[2] == 0x2F && testBytes[1] == 0xB5 && testBytes[0] == 0x28 {
-				return "application/zstd", nil
-			}
-		}
-		return contentType, nil
-	}
-	return blankType, nil
-}
